@@ -41,8 +41,8 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>2.3.2</version>
 				<configuration>
-					<source>1.6</source>
-					<target>1.6</target>
+					<source>13</source>
+					<target>13</target>
 				</configuration>
 			</plugin>
                         <plugin>
