@@ -27,34 +27,36 @@ package hu.mta.sztaki.lpds.cloud.simulator.energy.powermodelling;
 import hu.mta.sztaki.lpds.cloud.simulator.Timed;
 
 import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 
 /**
  * Represents a particular power state of a resource spreader
  * 
- * @author "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems, MTA SZTAKI (c) 2014"
+ * @author "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems,
+ *         MTA SZTAKI (c) 2014"
  *
  */
 public class PowerState {
 	/**
 	 * By sub-classing this class one can define arbitrary consumption models.
 	 * 
-	 * @author "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems, MTA SZTAKI (c) 2014"
+	 * @author "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems,
+	 *         MTA SZTAKI (c) 2014"
 	 *
 	 */
 	public static abstract class ConsumptionModel {
 		/**
-		 * backlink to the power state with useful data to determine the
-		 * consumption model
+		 * backlink to the power state with useful data to determine the consumption
+		 * model
 		 */
 		protected final PowerState myPowerState = null;
 
 		/**
-		 * calculates the instantaneous power draw of a resource spreader under
-		 * a particular load.
+		 * calculates the instantaneous power draw of a resource spreader under a
+		 * particular load.
 		 * 
-		 * @param load
-		 *            the load in the range of [0..1] of the resource spreader
+		 * @param load the load in the range of [0..1] of the resource spreader
 		 * @return the estimated instantaneous power draw of the spreader in W.
 		 */
 		protected abstract double evaluateConsumption(final double load);
@@ -66,17 +68,17 @@ public class PowerState {
 	 * characteristic change (like going from turboboost-&gt;using all cpu cores in
 	 * modern processors)
 	 * 
-	 * @author "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems, MTA SZTAKI (c) 2014"
+	 * @author "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems,
+	 *         MTA SZTAKI (c) 2014"
 	 *
 	 */
 	public interface PowerCharacteristicsChange {
 		/**
-		 * this function is called when there is a powerstate characteristic
-		 * change on the powerstate onMe.
+		 * this function is called when there is a powerstate characteristic change on
+		 * the powerstate onMe.
 		 * 
-		 * @param onMe
-		 *            with this parameter the change event propagates the power
-		 *            state for which the change is happening
+		 * @param onMe with this parameter the change event propagates the power state
+		 *             for which the change is happening
 		 */
 		void prePowerChangeEvent(final PowerState onMe);
 	}
@@ -86,8 +88,8 @@ public class PowerState {
 	 */
 	private double minConsumption;
 	/**
-	 * The power draw range in Watts (the maximum possible offset from minimum
-	 * power draw)
+	 * The power draw range in Watts (the maximum possible offset from minimum power
+	 * draw)
 	 */
 	private double consumptionRange;
 	/**
@@ -96,50 +98,49 @@ public class PowerState {
 	 */
 	private long pastNotification = -1;
 	/**
-	 * The current consumption model (which transforms the load to wattage based
-	 * on the power state's charactheristics (e.g., mincons/range)
+	 * The current consumption model (which transforms the load to wattage based on
+	 * the power state's charactheristics (e.g., mincons/range)
 	 */
 	private final ConsumptionModel model;
 	/**
 	 * the list of those objects who are prepared to handle power state changes
 	 */
-	private final ArrayList<PowerCharacteristicsChange> listeners = new ArrayList<PowerState.PowerCharacteristicsChange>();
+	private final ArrayList<PowerCharacteristicsChange> listeners = new ArrayList<>();
 
 	/**
 	 * Allow the creation of a new power state object with initial power state
 	 * characteristics.
 	 * 
-	 * @param minConsumption
-	 *            The minimum consumption that could be reported by this
-	 *            powerstate (e.g. idle power if we define a Physical Machine's
-	 *            powerstate) - in watts
-	 * @param consumptionRange
-	 *            The maximum offset of the consumption that could be still
-	 *            reported by this power state object. (e.g., the max-idle power
-	 *            if we define a Physical Machine) - in watts
-	 * @param modelclass
-	 *            Defines the consumption model which transforms a load value
-	 *            and is capable of outputting arbitrary instantaneous power
-	 *            draw estimates between minConsumption and
-	 *            minConsumption+consumptionRange
-	 * @throws InstantiationException
-	 *             if the modelclass is not existent
-	 * @throws IllegalAccessException
-	 *             if the modelclass is not accessible
-	 * @throws NoSuchFieldException
-	 *             if the modelclass does not have a myPowerState field (this is
-	 *             most likely impossible as all subclasses will have that field
-	 *             because of the baseclass)
-	 * @throws SecurityException
-	 *             if visibility changes are not possible to do on the
-	 *             myPowerState field
+	 * @param minConsumption   The minimum consumption that could be reported by
+	 *                         this powerstate (e.g. idle power if we define a
+	 *                         Physical Machine's powerstate) - in watts
+	 * @param consumptionRange The maximum offset of the consumption that could be
+	 *                         still reported by this power state object. (e.g., the
+	 *                         max-idle power if we define a Physical Machine) - in
+	 *                         watts
+	 * @param modelclass       Defines the consumption model which transforms a load
+	 *                         value and is capable of outputting arbitrary
+	 *                         instantaneous power draw estimates between
+	 *                         minConsumption and minConsumption+consumptionRange
+	 * @throws InstantiationException    if the modelclass is not existent
+	 * @throws IllegalAccessException    if the modelclass is not accessible
+	 * @throws NoSuchFieldException      if the modelclass does not have a
+	 *                                   myPowerState field (this is most likely
+	 *                                   impossible as all subclasses will have that
+	 *                                   field because of the baseclass)
+	 * @throws SecurityException         if visibility changes are not possible to
+	 *                                   do on the myPowerState field
+	 * @throws NoSuchMethodException
+	 * @throws InvocationTargetException
+	 * @throws IllegalArgumentException
 	 */
 	public PowerState(final double minConsumption, final double consumptionRange,
 			Class<? extends ConsumptionModel> modelclass)
-					throws InstantiationException, IllegalAccessException, NoSuchFieldException, SecurityException {
+			throws InstantiationException, IllegalAccessException, NoSuchFieldException, SecurityException,
+			IllegalArgumentException, InvocationTargetException, NoSuchMethodException {
 		this.minConsumption = minConsumption;
 		this.consumptionRange = consumptionRange;
-		model = modelclass.newInstance();
+		model = modelclass.getDeclaredConstructor(null).newInstance(null);
 		Field f = ConsumptionModel.class.getDeclaredField("myPowerState");
 		f.setAccessible(true);
 		f.set(model, this);
@@ -147,13 +148,12 @@ public class PowerState {
 	}
 
 	/**
-	 * determines the current power draw of the system given that it has the
-	 * load specified in the parameter
+	 * determines the current power draw of the system given that it has the load
+	 * specified in the parameter
 	 * 
-	 * @param load
-	 *            for which load value should we use the consumption model to
-	 *            determine the instantaneous power draw of the system. the load
-	 *            value is expected to be within the range of [0..1]
+	 * @param load for which load value should we use the consumption model to
+	 *             determine the instantaneous power draw of the system. the load
+	 *             value is expected to be within the range of [0..1]
 	 * @return the estimated instantaneous power draw value in watts
 	 */
 	public double getCurrentPower(final double load) {
@@ -182,8 +182,8 @@ public class PowerState {
 	}
 
 	/**
-	 * used to send out notifications before any of the power state
-	 * characteristics change
+	 * used to send out notifications before any of the power state characteristics
+	 * change
 	 */
 	private void notifyCharacteristisListeners() {
 		final long currentTime = Timed.getFireCount();
@@ -196,12 +196,11 @@ public class PowerState {
 	}
 
 	/**
-	 * Allows to set the minimum consumption for this power state (good for DVFS
-	 * and similar behavior). Before the change the power state change listeners
-	 * are notified.
+	 * Allows to set the minimum consumption for this power state (good for DVFS and
+	 * similar behavior). Before the change the power state change listeners are
+	 * notified.
 	 * 
-	 * @param minConsumption
-	 *            the new consumption value to be set in W
+	 * @param minConsumption the new consumption value to be set in W
 	 */
 	public void setMinConsumption(final double minConsumption) {
 		notifyCharacteristisListeners();
@@ -209,12 +208,11 @@ public class PowerState {
 	}
 
 	/**
-	 * Allows to set the consumption range for this power state (good for DVFS
-	 * and similar behavior). Before the change the power state change listeners
-	 * are notified.
+	 * Allows to set the consumption range for this power state (good for DVFS and
+	 * similar behavior). Before the change the power state change listeners are
+	 * notified.
 	 * 
-	 * @param cr
-	 *            the new consumption range value to be set in W
+	 * @param cr the new consumption range value to be set in W
 	 */
 	public void setConsumptionRange(final double cr) {
 		notifyCharacteristisListeners();
@@ -222,11 +220,10 @@ public class PowerState {
 	}
 
 	/**
-	 * allows users of this powerstate object to get notifications about power
-	 * state characteristics changes
+	 * allows users of this powerstate object to get notifications about power state
+	 * characteristics changes
 	 * 
-	 * @param listener
-	 *            the object where the events should be sent
+	 * @param listener the object where the events should be sent
 	 */
 	public void subscribePowerCharacteristicsChanges(PowerCharacteristicsChange listener) {
 		listeners.add(listener);
@@ -236,16 +233,15 @@ public class PowerState {
 	 * if some pary gets uninterested in power state changes this is the way to
 	 * cancel event notifications about them
 	 * 
-	 * @param listener
-	 *            the object which no longer needs notification event
+	 * @param listener the object which no longer needs notification event
 	 */
 	public void unsubscribePowerCharacteristicsChanges(PowerCharacteristicsChange listener) {
 		listeners.remove(listener);
 	}
 
 	/**
-	 * A convenient and compact way to represent the main characteristics of
-	 * this power state. Good for debugging and tracing
+	 * A convenient and compact way to represent the main characteristics of this
+	 * power state. Good for debugging and tracing
 	 */
 	@Override
 	public String toString() {
