@@ -24,20 +24,19 @@
  */
 package hu.mta.sztaki.lpds.cloud.simulator.iaas.pmscheduling;
 
+import java.util.HashMap;
+import java.util.List;
+
 import hu.mta.sztaki.lpds.cloud.simulator.DeferredEvent;
 import hu.mta.sztaki.lpds.cloud.simulator.iaas.IaaSService;
 import hu.mta.sztaki.lpds.cloud.simulator.iaas.PhysicalMachine;
 import hu.mta.sztaki.lpds.cloud.simulator.iaas.PhysicalMachine.State;
 import hu.mta.sztaki.lpds.cloud.simulator.iaas.VMManager;
-import hu.mta.sztaki.lpds.cloud.simulator.iaas.VMManager.CapacityChangeEvent;
 import hu.mta.sztaki.lpds.cloud.simulator.iaas.VMManager.VMManagementException;
 import hu.mta.sztaki.lpds.cloud.simulator.iaas.constraints.AlterableResourceConstraints;
 import hu.mta.sztaki.lpds.cloud.simulator.iaas.constraints.ResourceConstraints;
 import hu.mta.sztaki.lpds.cloud.simulator.iaas.vmscheduling.Scheduler;
 import hu.mta.sztaki.lpds.cloud.simulator.io.NetworkNode.NetworkException;
-
-import java.util.HashMap;
-import java.util.List;
 
 /**
  * A reactive PM controller that increases/decreases the powered on pm set on
@@ -45,8 +44,9 @@ import java.util.List;
  * changing the machine set size by one, thus it is not applicable in cases when
  * there are rapid demand changes on the infrastructure.
  * 
- * @author "Gabor Kecskemeti, Distributed and Parallel Systems Group, University of Innsbruck (c) 2013"
- *         "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems, MTA SZTAKI (c) 2012"
+ * @author "Gabor Kecskemeti, Distributed and Parallel Systems Group, University
+ *         of Innsbruck (c) 2013" "Gabor Kecskemeti, Laboratory of Parallel and
+ *         Distributed Systems, MTA SZTAKI (c) 2012"
  */
 public class SchedulingDependentMachines extends PhysicalMachineController {
 
@@ -57,8 +57,9 @@ public class SchedulingDependentMachines extends PhysicalMachineController {
 	 * PM's free resources reach its complete resource set and there are no VM
 	 * requests queuing at the IaaS service's VM scheduler.
 	 * 
-	 * @author "Gabor Kecskemeti, Distributed and Parallel Systems Group, University of Innsbruck (c) 2013"
-	 *         "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems, MTA SZTAKI (c) 2012"
+	 * @author "Gabor Kecskemeti, Distributed and Parallel Systems Group, University
+	 *         of Innsbruck (c) 2013" "Gabor Kecskemeti, Laboratory of Parallel and
+	 *         Distributed Systems, MTA SZTAKI (c) 2012"
 	 */
 	private class CapacityChangeManager
 			implements VMManager.CapacityChangeEvent<ResourceConstraints>, PhysicalMachine.StateChangeListener {
@@ -70,13 +71,12 @@ public class SchedulingDependentMachines extends PhysicalMachineController {
 		final PhysicalMachine observed;
 
 		/**
-		 * This constructor is expected to be used once a PM is registered to
-		 * the parent IaaSService. From that point on the newly registered PM
-		 * will be considered for switching off/turning on.
+		 * This constructor is expected to be used once a PM is registered to the parent
+		 * IaaSService. From that point on the newly registered PM will be considered
+		 * for switching off/turning on.
 		 * 
-		 * @param pm
-		 *            the PM to be managed to follow the demand patterns of the
-		 *            VM requests to the IaaSService.
+		 * @param pm the PM to be managed to follow the demand patterns of the VM
+		 *           requests to the IaaSService.
 		 */
 		public CapacityChangeManager(final PhysicalMachine pm) {
 			observed = pm;
@@ -85,9 +85,9 @@ public class SchedulingDependentMachines extends PhysicalMachineController {
 		}
 
 		/**
-		 * if the PM gets dropped from the parent IaaSService then we no longer
-		 * need to control its behavior. So the class's active behavior is
-		 * disabled by calling this function.
+		 * if the PM gets dropped from the parent IaaSService then we no longer need to
+		 * control its behavior. So the class's active behavior is disabled by calling
+		 * this function.
 		 */
 		private void cancelEvents() {
 			observed.unsubscribeFromIncreasingFreeCapacityChanges(this);
@@ -108,11 +108,10 @@ public class SchedulingDependentMachines extends PhysicalMachineController {
 		}
 
 		/**
-		 * This function is called when the observed PM has newly freed up
-		 * resources. It ensures that the PM is switched off if the PM is
-		 * completely free and there are no more VMs queuing at the VM scheduler
-		 * (i.e., there will be no chance to receive a new VM for the capacities
-		 * of this PM).
+		 * This function is called when the observed PM has newly freed up resources. It
+		 * ensures that the PM is switched off if the PM is completely free and there
+		 * are no more VMs queuing at the VM scheduler (i.e., there will be no chance to
+		 * receive a new VM for the capacities of this PM).
 		 */
 		@Override
 		public void capacityChanged(final ResourceConstraints newCapacity,
@@ -125,11 +124,11 @@ public class SchedulingDependentMachines extends PhysicalMachineController {
 		}
 
 		/**
-		 * This function is called when the PM's power state changes. This event
-		 * handler manages situations when the PM is turned on but there are no
-		 * longer tasks for it. Also, it initiates a new PM's switchon if the
-		 * newly switched on machine will not be enough to host the queued VM
-		 * requests found at the VM scheduler of the parent IaaS service.
+		 * This function is called when the PM's power state changes. This event handler
+		 * manages situations when the PM is turned on but there are no longer tasks for
+		 * it. Also, it initiates a new PM's switchon if the newly switched on machine
+		 * will not be enough to host the queued VM requests found at the VM scheduler
+		 * of the parent IaaS service.
 		 */
 		@Override
 		public void stateChanged(PhysicalMachine pm, State oldState, State newState) {
@@ -167,53 +166,48 @@ public class SchedulingDependentMachines extends PhysicalMachineController {
 	/**
 	 * this map lists all the currently controlled PMs and their controllers.
 	 */
-	private HashMap<PhysicalMachine, CapacityChangeManager> capacityManagers = new HashMap<PhysicalMachine, CapacityChangeManager>();
+	private HashMap<PhysicalMachine, CapacityChangeManager> capacityManagers = new HashMap<>();
 	/**
-	 * ensures that we only have a single machine switching on at a time and
-	 * shows what is the actual machine that is switching on.
+	 * ensures that we only have a single machine switching on at a time and shows
+	 * what is the actual machine that is switching on.
 	 */
 	private PhysicalMachine currentlyStartingPM = null;
 
 	/**
-	 * Constructs the scheduler and passes the parent IaaSService to the
-	 * superclass.
+	 * Constructs the scheduler and passes the parent IaaSService to the superclass.
 	 * 
-	 * @param parent
-	 *            the IaaSService to serve
+	 * @param parent the IaaSService to serve
 	 */
 	public SchedulingDependentMachines(final IaaSService parent) {
 		super(parent);
 	}
 
 	/**
-	 * Defines to do the following when a new host is (de)registered to the
-	 * parent IaaSService:
+	 * Defines to do the following when a new host is (de)registered to the parent
+	 * IaaSService:
 	 * <ul>
-	 * <li>if the current event is a registration event then the function
-	 * creates and locally registers a new capacity change manager for the newly
-	 * registered pms
-	 * <li>if the current event is a deregistration event then the function
-	 * cancels the capacity management for all deregistered pms
+	 * <li>if the current event is a registration event then the function creates
+	 * and locally registers a new capacity change manager for the newly registered
+	 * pms
+	 * <li>if the current event is a deregistration event then the function cancels
+	 * the capacity management for all deregistered pms
 	 * </ul>
 	 */
 	@Override
 	protected VMManager.CapacityChangeEvent<PhysicalMachine> getHostRegEvent() {
-		return new CapacityChangeEvent<PhysicalMachine>() {
-			@Override
-			public void capacityChanged(final ResourceConstraints newCapacity, final List<PhysicalMachine> alteredPMs) {
-				final boolean newRegistration = parent.isRegisteredHost(alteredPMs.get(0));
-				final int pmNum = alteredPMs.size();
-				if (newRegistration) {
-					// Management of capacity increase
-					for (int i = pmNum - 1; i >= 0; i--) {
-						final PhysicalMachine pm = alteredPMs.get(i);
-						capacityManagers.put(pm, new CapacityChangeManager(pm));
-					}
-				} else {
-					// Management of capacity decrease
-					for (int i = pmNum - 1; i >= 0; i--) {
-						capacityManagers.remove(alteredPMs.get(i)).cancelEvents();
-					}
+		return (final ResourceConstraints newCapacity, final List<PhysicalMachine> alteredPMs) -> {
+			final boolean newRegistration = parent.isRegisteredHost(alteredPMs.get(0));
+			final int pmNum = alteredPMs.size();
+			if (newRegistration) {
+				// Management of capacity increase
+				for (int i = pmNum - 1; i >= 0; i--) {
+					final PhysicalMachine pm = alteredPMs.get(i);
+					capacityManagers.put(pm, new CapacityChangeManager(pm));
+				}
+			} else {
+				// Management of capacity decrease
+				for (int i = pmNum - 1; i >= 0; i--) {
+					capacityManagers.remove(alteredPMs.get(i)).cancelEvents();
 				}
 			}
 		};
@@ -222,32 +216,28 @@ public class SchedulingDependentMachines extends PhysicalMachineController {
 	/**
 	 * Defines to do the following when VM requests arrive:
 	 * <ol>
-	 * <li>check if there is already a PM under preparation (if there is one
-	 * then does nothing)
-	 * <li>if there are no PM that is currently prepared then starts to prepare
-	 * one for accepting VM requests.
+	 * <li>check if there is already a PM under preparation (if there is one then
+	 * does nothing)
+	 * <li>if there are no PM that is currently prepared then starts to prepare one
+	 * for accepting VM requests.
 	 * </ol>
 	 */
 	@Override
 	protected Scheduler.QueueingEvent getQueueingEvent() {
-		return new Scheduler.QueueingEvent() {
-			@Override
-			public void queueingStarted() {
-				if (currentlyStartingPM == null
-						|| PhysicalMachine.State.RUNNING.equals(currentlyStartingPM.getState())) {
-					// If there are no machines under their startup procedure,
-					// or the currently started up machine is already running
-					// and we still receive the queueingstarted event
-					turnOnAMachine();
-				}
+		return () -> {
+			if (currentlyStartingPM == null || PhysicalMachine.State.RUNNING.equals(currentlyStartingPM.getState())) {
+				// If there are no machines under their startup procedure,
+				// or the currently started up machine is already running
+				// and we still receive the queueingstarted event
+				turnOnAMachine();
 			}
 		};
 	}
 
 	/**
-	 * switches on a not yet switched on machine from the parent IaaS's PM set.
-	 * if there are no more machines in the IaaS that can be turned on then the
-	 * calling of this function is ignored.
+	 * switches on a not yet switched on machine from the parent IaaS's PM set. if
+	 * there are no more machines in the IaaS that can be turned on then the calling
+	 * of this function is ignored.
 	 */
 	protected void turnOnAMachine() {
 		final int pmsize = parent.machines.size();
