@@ -37,7 +37,8 @@ import hu.mta.sztaki.lpds.cloud.simulator.notifications.StateDependentEventHandl
  * implements a notification handler for sending out notifications about power
  * state changes in resource spreaders
  * 
- * @author "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems, MTA SZTAKI (c) 2015"
+ * @author "Gabor Kecskemeti, Laboratory of Parallel and Distributed Systems,
+ *         MTA SZTAKI (c) 2015"
  *
  */
 
@@ -45,8 +46,8 @@ public class PowerStateChangeNotificationHandler
 		implements SingleNotificationHandler<PowerBehaviorChangeListener, Pair<ResourceSpreader, PowerState>> {
 
 	/**
-	 * the single object that will handle all notification operations on the
-	 * same way
+	 * the single object that will handle all notification operations on the same
+	 * way
 	 */
 	private static final PowerStateChangeNotificationHandler handlerSingleton = new PowerStateChangeNotificationHandler();
 
@@ -59,25 +60,23 @@ public class PowerStateChangeNotificationHandler
 	}
 
 	/**
-	 * gets the event handler that will manage the notification subscriptions
-	 * for the particular resource spreader object that asked for the handler.
+	 * gets the event handler that will manage the notification subscriptions for
+	 * the particular resource spreader object that asked for the handler.
 	 * 
 	 * @return the eventh handler
 	 */
 	public static StateDependentEventHandler<PowerBehaviorChangeListener, Pair<ResourceSpreader, PowerState>> getHandlerInstance() {
-		return new StateDependentEventHandler<PowerBehaviorChangeListener, Pair<ResourceSpreader, PowerState>>(
-				handlerSingleton);
+		return new StateDependentEventHandler<>(handlerSingleton);
 	}
 
 	/**
 	 * The event handling mechanism for power state change notifications about
 	 * resource spreaders
 	 * 
-	 * @param onObject
-	 *            The listener to send the event to
-	 * @param newPowerBehavior
-	 *            a data pair containing the resource spreader that changed its
-	 *            state and the power state that the spreader just switches to.
+	 * @param onObject         The listener to send the event to
+	 * @param newPowerBehavior a data pair containing the resource spreader that
+	 *                         changed its state and the power state that the
+	 *                         spreader just switches to.
 	 */
 	@Override
 	public void sendNotification(final PowerBehaviorChangeListener onObject,
